Organize  dir
