+++ alt_dev/run_alt.py
@@ -33,7 +33,7 @@ import pandas as pd
 # NREL dev API key, for weather and data files and site information
 from pathlib import Path
 from hopp.sites import make_circular_site, make_irregular_site, SiteInfo, locations
-from hopp.keys import set_nrel_key_dot_env
+from hopp.utils.keys import set_nrel_key_dot_env
 set_nrel_key_dot_env()
 
 from functools import partial
