+++ examples/H2_Analysis/h2_setup_optimize.py
@@ -8,8 +8,8 @@ import pandas as pd
 from hopp.sites import SiteInfo
 from hopp.sites import flatirons_site as sample_site
 from hopp.hybrid_simulation import HybridSimulation
-from hopp.log import hybrid_logger as logger
-from hopp.keys import set_developer_nrel_gov_key
+from hopp.utils.log import hybrid_logger as logger
+from hopp.utils.keys import set_developer_nrel_gov_key
 from examples.H2_Analysis.simple_dispatch import SimpleDispatch
 from examples.H2_Analysis.run_h2_PEM import run_h2_PEM
 import numpy as np
