+++ examples/analysis/multi_location.py
@@ -20,8 +20,8 @@ from pathlib import Path
 from dotenv import load_dotenv
 from itertools import repeat
 
-from hopp.keys import set_developer_nrel_gov_key
-from hopp.log import analysis_logger as logger
+from hopp.utils.keys import set_developer_nrel_gov_key
+from hopp.utils.log import analysis_logger as logger
 from hopp.sites import SiteInfo
 from hopp.sites import flatirons_site as sample_site
 from hopp.hybrid_simulation import HybridSimulation
