+++ examples/analysis/single_location.py
@@ -17,8 +17,8 @@ import operator
 from pathlib import Path
 from itertools import repeat
 
-from hopp.keys import set_nrel_key_dot_env
-from hopp.log import analysis_logger as logger
+from hopp.utils.keys import set_nrel_key_dot_env
+from hopp.utils.log import analysis_logger as logger
 from hopp.sites import SiteInfo
 from hopp.sites import flatirons_site as sample_site
 from hopp.hybrid_simulation import HybridSimulation
