+++ examples/flicker.py
@@ -4,7 +4,7 @@ import matplotlib.pyplot as plt
 
 from hopp.layout.flicker_mismatch_grid import FlickerMismatch, FlickerMismatchGrid
 from hopp.layout.flicker_data.plot_flicker import plot_contour
-from hopp.keys import set_developer_nrel_gov_key
+from hopp.utils.keys import set_developer_nrel_gov_key
 
 
 # Set API key
