+++ examples/optimization/layout_opt/hybrid_run.py
@@ -35,9 +35,9 @@ from tools.optimization import (
     DataRecorder
     )
 from hopp.sites import make_circular_site, make_irregular_site, SiteInfo
-from hopp.log import opt_logger as logger
+from hopp.utils.log import opt_logger as logger
 from hopp.sites import locations
-from hopp.keys import set_developer_nrel_gov_key
+from hopp.utils.keys import set_developer_nrel_gov_key
 from hopp.layout.plot_tools import *
 
 from parametrized_optimization_driver import ParametrizedOptimizationDriver
