+++ examples/simulate_hybrid.py
@@ -3,8 +3,8 @@ from pathlib import Path
 
 from hopp.sites import SiteInfo, flatirons_site
 from hopp.hybrid_simulation import HybridSimulation
-from hopp.log import hybrid_logger as logger
-from hopp.keys import set_nrel_key_dot_env
+from hopp.utils.log import hybrid_logger as logger
+from hopp.utils.keys import set_nrel_key_dot_env
 
 examples_dir = Path(__file__).parent.absolute()
