+++ hopp/power_source.py
@@ -5,7 +5,7 @@ import PySAM.Singleowner as Singleowner
 import PySAM.Pvsamv1 as Pvsamv1
 import pandas as pd
 from tools.utils import flatten_dict, array_not_scalar
-from hopp.log import hybrid_logger as logger
+from hopp.utils.log import hybrid_logger as logger
 from hopp.dispatch.power_sources.power_source_dispatch import PowerSourceDispatch
