+++ hopp/reopt.py
@@ -6,11 +6,11 @@ import time
 
 from typing import Sequence
 
-from hopp.pv_source import *
-from hopp.wind_source import WindPlant
+from hopp.pv_plant import *
+from hopp.wind_plant import WindPlant
 from hopp.battery import Battery
-from hopp.log import hybrid_logger as logger
-from hopp.keys import get_developer_nrel_gov_key
+from hopp.utils.log import hybrid_logger as logger
+from hopp.utils.keys import get_developer_nrel_gov_key
 from hopp.utility_rate import UtilityRate
 
 import urllib3
