+++ hopp/utility_rate.py
@@ -2,7 +2,7 @@ import json
 import os
 import requests
 
-from hopp.keys import get_developer_nrel_gov_key
+from hopp.utils.keys import get_developer_nrel_gov_key
 
 
 class UtilityRate:
