+++ tests/analysis/test_hopp_for_h2.py
@@ -10,7 +10,7 @@ from tools.resource.resource_tools import *
 from hopp.sites import flatirons_site as sample_site
 from examples.H2_Analysis.hopp_for_h2 import hopp_for_h2
 from hopp.sites import SiteInfo
-from hopp.keys import set_developer_nrel_gov_key
+from hopp.utils.keys import set_developer_nrel_gov_key
 
 # Set API key
 load_dotenv()
