+++ tests/hopp/test_hybrid.py
@@ -10,7 +10,7 @@ from examples.Detailed_PV_Layout.detailed_pv_layout import DetailedPVParameters,
 from examples.Detailed_PV_Layout.detailed_pv_config import PVLayoutConfig
 import PySAM.Singleowner as Singleowner
 from hopp.grid import Grid
-from hopp.keys import set_nrel_key_dot_env
+from hopp.utils.keys import set_nrel_key_dot_env
 from hopp.layout.pv_design_utils import size_electrical_parameters
 from copy import deepcopy
 import numpy as np
